+++ aitune/torch/task/profiling/config.py
@@ -5,7 +5,6 @@
 from collections.abc import Generator
 from dataclasses import dataclass
 
-from aitune.torch.config import DEFAULT_THROUGHPUT_BACKOFF_LIMIT, DEFAULT_THROUGHPUT_CUTOFF_THRESHOLD
 from aitune.torch.task.profiling.measuring_stop_strategy import MeasuringStopStrategy, NumStepsMeasuringStopStrategy
 from aitune.torch.task.profiling.measuring_strategy import MeasuringStrategy, ModelExecutionTimeMeasuringStrategy
 from aitune.torch.task.profiling.profiling_stop_strategy import (
@@ -22,9 +21,10 @@ class ProfilingConfig:
 
     Attributes:
         batch_sizes: List of batch sizes to profile.
-        measuring_strategy: Strategy to measure the model. Default is ModelExecutionTimeMeasuringStrategy().
-        measurement_stop_strategy: Strategy to stop collecting measurements. Default is NumStepsMeasuringStopStrategy(num_steps=10).
-        profiling_stop_strategy: Strategy to stop profiling. Default is ThroughputSaturatedProfilingStopStrategy(throughput_cutoff_threshold=0.05).
+        batching: Whether profiling should batch samples together.
+        measuring_strategy: Strategy to measure the model.
+        measurement_stop_strategy: Strategy to stop collecting measurements.
+        profiling_stop_strategy: Strategy to stop profiling.
     """
 
     batch_sizes: list[int] | Generator[int, None, None]
@@ -36,6 +36,9 @@ class ProfilingConfig:
 
     def __post_init__(self):
         """Post-init."""
+        if isinstance(self.batch_sizes, Generator):
+            self.batch_sizes = list(self.batch_sizes)
+
         if not self.batch_sizes:
             raise ValueError("batch_sizes must be provided")
 
@@ -43,10 +46,7 @@ class ProfilingConfig:
             self.measuring_strategy = ModelExecutionTimeMeasuringStrategy()
 
         if self.measurement_stop_strategy is None:
-            self.measurement_stop_strategy = NumStepsMeasuringStopStrategy(num_steps=10)
+            self.measurement_stop_strategy = NumStepsMeasuringStopStrategy()
 
         if self.profiling_stop_strategy is None:
-            self.profiling_stop_strategy = ThroughputSaturatedProfilingStopStrategy(
-                throughput_cutoff_threshold=DEFAULT_THROUGHPUT_CUTOFF_THRESHOLD,
-                throughput_backoff_limit=DEFAULT_THROUGHPUT_BACKOFF_LIMIT,
-            )
+            self.profiling_stop_strategy = ThroughputSaturatedProfilingStopStrategy()