+++ tests/unit/torch/tune_strategy/test_to_config_dict.py
@@ -51,6 +51,6 @@ def test_max_throughput_strategy_to_json_dict():
 
     assert result == {
         "backends": ["BackendA", "BackendB"],
-        "measurement_stop_strategy": "StableWindowMeasuringStopStrategy",
+        "measurement_stop_strategy": "NumStepsMeasuringStopStrategy",
         "profiling_stop_strategy": "ThroughputSaturatedProfilingStopStrategy",
     }